requirements-test.txt
pytest==6.2.5
pytest-homeassistant-custom-component>=0.5.0
requests-mock==1.9.2
pytest-cov==2.12.1